# docs/api/openapi.yaml
...
openapi: 3.1.0
info:
  title: ComplyTime Platform API
  version: 0.1.0
  description: >
    REST surface served by `cmd/gateway`. Many `/api/*` routes register only when PostgreSQL
    is configured; `/api/config` is always available. After Google OAuth, the encrypted
    `studio_session` cookie authenticates `/api/*` requests (unless disabled). Optional
    `Authorization: Bearer <STUDIO_API_TOKEN>` bypasses the cookie for scripting.
    Security schemes below model bearer token and OAuth; cookie sessions are established via `/auth/*`.
servers:
  - url: http://localhost:8080
    description: Local gateway default (PORT=8080)
security:
  - bearerAuth: []
  - oauth2: [openid, email, profile]
...
paths:
...
  /api/policies:
    get:
      tags: [Policies]
      operationId: listPolicies
      summary: List imported policies
      responses:
        '200':
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Policy'
        '500':
          description: Internal error
...
  /api/ingest:
    post:
      tags: [Ingest]
      operationId: ingestArtifact
      summary: Async ingest any Gemara artifact
      description: >-
        Accepts raw Gemara YAML (any artifact type — EvaluationLog,
        EnforcementLog, Policy, ControlCatalog, ThreatCatalog,
        RiskCatalog, GuidanceCatalog, MappingDocument). Returns 202
        with a job_id. Poll GET /api/ingest/jobs/{job_id} for status.
        See ADR #0034.
      requestBody:
        required: true
        content:
          application/x-yaml:
            schema:
              type: string
          application/json:
            schema:
              type: string
      responses:
        '202':
          description: Accepted — ingest queued
          content:
            application/json:
              schema:
                type: object
                properties:
                  job_id:
                    type: string
                  status:
                    type: string
                    example: pending
        '400':
          description: Empty body
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorsListResponse'
  /api/ingest/jobs/{job_id}:
    get:
      tags: [Ingest]
      operationId: getIngestJobStatus
      summary: Poll async ingest job status
      parameters:
        - name: job_id
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Job status
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/IngestJobStatus'
        '404':
          description: Job not found (expired or never existed)
...
  /api/audit-logs:
    get:
      tags: [AuditLogs]
      operationId: listAuditLogs
      summary: List audit logs for a policy
      parameters:
        - name: policy_id
          in: query
          required: true
          schema:
            type: string
        - name: start
          in: query
          description: Date YYYY-MM-DD
          schema:
            type: string
        - name: end
          in: query
          schema:
            type: string
        - name: limit
          in: query
          schema:
            type: integer
      responses:
        '200':
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/AuditLog'
        '400':
          description: Missing policy_id
    post:
      tags: [AuditLogs]
      operationId: createAuditLog
      summary: Store finalized audit log YAML
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateAuditLogRequest'
      responses:
        '201':
          content:
            application/json:
              schema:
                type: object
                properties:
                  status:
                    type: string
                  audit_id:
                    type: string
        '400':
          description: Invalid YAML or JSON
...
  /api/bootstrap:
    post:
      tags: [Users]
      operationId: bootstrapAdmin
      summary: Promote caller to admin when no admins exist
      responses:
        '200':
          content:
            application/json:
              schema:
                type: object
                properties:
                  email:
                    type: string
                  role:
                    type: string
        '401':
          description: Unauthenticated
        '409':
          description: Admin already exists
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      description: Shared secret from `STUDIO_API_TOKEN`; treated as admin-capable service identity when configured.
    oauth2:
      type: oauth2
      description: >
        Google OAuth 2.0 login flow used to mint the encrypted `studio_session` cookie.
        API requests authenticate via that cookie unless `bearerAuth` matches.
      flows:
        authorizationCode:
          authorizationUrl: https://accounts.google.com/o/oauth2/v2/auth
          tokenUrl: https://oauth2.googleapis.com/token
          scopes:
            openid: OpenID Connect
            email: Email
            profile: Profile
...
  schemas:
...
    ErrorsListResponse:
      type: object
      properties:
        errors:
          type: array
          items:
            type: string
    Policy:
      type: object
      properties:
        policy_id:
          type: string
        title:
          type: string
        version:
          type: string
        oci_reference:
          type: string
        content:
          type: string
        imported_at:
          type: string
          format: date-time
        imported_by:
          type: string
...
    AuditLog:
      type: object
      properties:
        audit_id:
          type: string
        policy_id:
          type: string
        audit_start:
          type: string
          format: date-time
        audit_end:
          type: string
          format: date-time
        framework:
          type: string
        created_at:
          type: string
          format: date-time
        created_by:
          type: string
        content:
          type: string
        summary:
          type: string
        model:
          type: string
        prompt_version:
          type: string
...
    CreateAuditLogRequest:
      type: object
      required: [policy_id, content]
      properties:
        policy_id:
          type: string
        content:
          type: string
        model:
          type: string
        prompt_version:
          type: string
...
    IngestJobStatus:
      type: object
      properties:
        job_id:
          type: string
        status:
          type: string
          enum: [pending, completed, failed]
        inserted:
          type: integer
          minimum: 0
          description: Evidence row count (EvaluationLog/EnforcementLog only)
        policy_id:
          type: string
          description: Derived policy ID (evidence only)
        artifact_id:
          type: string
          description: Stored artifact ID (Policy, Catalog, Mapping)
        artifact_type:
          type: string
          description: Detected Gemara artifact type
        error:
          type: string
        created_at:
          type: string
          format: date-time
        updated_at:
          type: string
          format: date-time
...